pyproject.toml
[project]
name = "cornell-app-dev"
version = "0.1.0"
description = "Project directory for the Cornell App Development Club ( Intro to Backend Development Assignments )"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "passlib>=1.7.4",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
]
